+++ CHANGES.md
@@ -57,6 +57,8 @@
 * Change increase the scope and number of non release group text that is identified and removed
 * Add a general config setting to allow adding incomplete show data
 * Change to throttle connection rate on thread initiation for adba library
+* Change default manage episodes selector to Snatched episodes if items exist else Wanted on Episode Status Manage page
+* Change snatched row colour on Episode Status Manage page to match colour used on the show details page
 
 [develop changelog]
 Enable Alpha Ratio again now that the secure login page over https is fixed